Code refactoring: moving constants and ApplyDefault from v1alpha1 to controllers.
Specially
- constants.go:
  constants was declared in v1alpha1 package but used by controllers package, thus moving
- utils.go:
  move some util functions to this file
- Fixes
  type ServiceName string is missing "=" before string, otherwise it becomes a typedef
  (we need is a type alias)
- Removed cdapmaster.go
  ApplyDefault() there should be together with controller, thus being moved.
